refactor(aggregator): rename `gcp_uploader` module to `cloud_uploader`
As its components are not tied to GCP but only the backend.
