+++ mithril-aggregator/src/file_uploaders/cloud_uploader/interface.rs
