+++ mithril-aggregator/src/file_uploaders/mod.rs
@@ -1,12 +1,12 @@
+mod cloud_uploader;
 mod dumb_uploader;
-mod gcp_uploader;
 mod interface;
 mod local_uploader;
 
-pub use dumb_uploader::*;
-pub use gcp_uploader::{
-    CloudRemotePath, GCloudBackendUploader, GcpBackendUploaderLegacy, GcpUploader,
+pub use cloud_uploader::{
+    CloudRemotePath, CloudUploader, GCloudBackendUploader, GcpBackendUploaderLegacy,
 };
+pub use dumb_uploader::*;
 pub use interface::{FileUploadRetryPolicy, FileUploader};
 pub use local_uploader::LocalUploader;
